Consolidate SqrlConfig tests into single test file and fix temp file handling
- Rename SqrlConfigEdgeCasesTest to SqrlConfigTest as the main test file
- Merge tests from SqrlConfigCommonsTest and ConfigurationTest into SqrlConfigTest
- Remove duplicate test files to reduce maintenance overhead
- Fix createTempFile method to use system temp directory instead of source directory
- Update PackageJsonSchemaTest imports to reference consolidated test file
- All 23 SqrlConfig tests pass with comprehensive coverage of interface and implementation
