+++ sqrl-tools/sqrl-config/src/test/java/com/datasqrl/config/SqrlConfigTest.java
@@ -17,19 +17,25 @@ package com.datasqrl.config;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.Assertions.fail;
 
 import com.datasqrl.error.CollectedException;
 import com.datasqrl.error.ErrorCollector;
+import com.datasqrl.error.ErrorPrefix;
+import com.datasqrl.error.ErrorPrinter;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
 import java.util.Map;
+import java.util.function.Consumer;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-class SqrlConfigEdgeCasesTest {
+class SqrlConfigTest {
 
   private ErrorCollector errors;
   private SqrlConfig config;
@@ -37,7 +43,7 @@ class SqrlConfigEdgeCasesTest {
 
   @BeforeEach
   void setUp() {
-    errors = ErrorCollector.root();
+    errors = new ErrorCollector(ErrorPrefix.ROOT);
     config = SqrlConfig.createCurrentVersion();
   }
 
@@ -248,6 +254,138 @@ class SqrlConfigEdgeCasesTest {
     assertThat(newConfig.getVersion()).isEqualTo(1);
   }
 
+  @Test
+  void givenNoPaths_whenCreatingConfigFromPackageJson_thenReturnsDefaults() {
+    var underTest = SqrlConfigCommons.fromFilesPackageJson(errors, List.of());
+
+    assertThat(underTest).isNotNull();
+    assertThat(underTest.getVersion()).isEqualTo(1);
+    assertThat(underTest.getEnabledEngines()).contains("vertx", "postgres", "kafka", "flink");
+    assertThat(underTest.getTestConfig()).isPresent();
+    assertThat(underTest.getEngines().getEngineConfig("flink")).isPresent();
+    assertThat(underTest.getScriptConfig().getGraphql()).isEmpty();
+    assertThat(underTest.getScriptConfig().getMainScript()).isEmpty();
+  }
+
+  @Test
+  void givenSinglePath_whenCreatingConfigFromPackageJson_thenOverridesDefaults() {
+    var underTest =
+        SqrlConfigCommons.fromFilesPackageJson(
+            errors, List.of(Path.of("src/test/resources/config/test-package.json")));
+
+    assertThat(underTest).isNotNull();
+    assertThat(underTest.getVersion()).isEqualTo(1);
+
+    assertThat(underTest.getEnabledEngines()).contains("test");
+
+    assertThat(underTest.getTestConfig()).isPresent();
+    assertThat(underTest.getEngines().getEngineConfig("flink")).isPresent();
+    assertThat(underTest.getScriptConfig().getGraphql()).isEmpty();
+    assertThat(underTest.getScriptConfig().getMainScript()).isEmpty();
+  }
+
+  @Test
+  void givenJsonConfigFile_whenLoadViaCommons_thenParsesCorrectly() {
+    SqrlConfig config1 =
+        SqrlConfigCommons.fromFiles(errors, Path.of("src/test/resources/config/config1.json"));
+    testConfig1(config1);
+  }
+
+  @Test
+  void givenLoadedConfig_whenWriteToFile_thenLoadsIdentically() {
+    var loadedConfig =
+        SqrlConfigCommons.fromFiles(errors, Path.of("src/test/resources/config/config1.json"));
+    var tempFile2 = createTempFile();
+    loadedConfig.toFile(tempFile2);
+    var reloadedConfig = SqrlConfigCommons.fromFiles(errors, tempFile2);
+    testConfig1(reloadedConfig);
+  }
+
+  @Test
+  void givenSubConfig_whenWriteToFile_thenLoadsCorrectSubset() {
+    var loadedConfig =
+        SqrlConfigCommons.fromFiles(errors, Path.of("src/test/resources/config/config1.json"));
+    var tempFile2 = createTempFile();
+    loadedConfig.getSubConfig("subConf").toFile(tempFile2, true);
+    var subConfig = SqrlConfigCommons.fromFiles(errors, tempFile2);
+    testSubConf(subConfig);
+  }
+
+  @Test
+  void givenSourceConfig_whenCopy_thenCopiesConfiguration() {
+    var other =
+        SqrlConfigCommons.fromFiles(errors, Path.of("src/test/resources/config/config1.json"))
+            .getSubConfig("subConf");
+    var newConf = SqrlConfig.createCurrentVersion();
+    newConf.copy(other);
+    testSubConf(newConf);
+  }
+
+  @Test
+  void givenNewConfig_whenSetPropertiesAndObjects_thenPersistsCorrectly() {
+    var newConf = SqrlConfig.createCurrentVersion();
+    newConf.setProperty("test", true);
+    var tc = new TestClass(9, "boat", List.of("x", "y", "z"));
+    newConf.getSubConfig("clazz").setProperties(tc);
+    assertThat(newConf.asBool("test").get()).isTrue();
+    assertThat(newConf.getSubConfig("clazz").allAs(TestClass.class).get().field3)
+        .isEqualTo(tc.field3);
+    var tempFile2 = createTempFile();
+    newConf.toFile(tempFile2, true);
+    var config2 = SqrlConfigCommons.fromFiles(errors, tempFile2);
+    assertThat(config2.asBool("test").get()).isTrue();
+    var tc2 = config2.getSubConfig("clazz").allAs(TestClass.class).get();
+    assertThat(tc2.field1).isEqualTo(tc.field1);
+    assertThat(tc2.field2).isEqualTo(tc.field2);
+    assertThat(tc2.field3).isEqualTo(tc.field3);
+  }
+
+  private void testConfig1(SqrlConfig config) {
+    assertThat(config.asInt("key2").get()).isEqualTo(5);
+    assertThat(config.asLong("key2").get()).isEqualTo(5L);
+    assertThat(config.asString("key1").get()).isEqualTo("value1");
+    assertThat(config.asBool("key3").get()).isTrue();
+    assertThat(config.asList("list", String.class).get()).isEqualTo(List.of("a", "b", "c"));
+    Map<String, TestClass> map = config.asMap("map", TestClass.class).get();
+    assertThat(map).hasSize(3);
+    assertThat(map.get("e2").field1).isEqualTo(7);
+    assertThat(map.get("e3").field2).isEqualTo("flip");
+    assertThat(map.get("e1").field3).isEqualTo(List.of("a", "b", "c"));
+    assertThat(config.getVersion()).isEqualTo(1);
+
+    var x1 = config.as("x1", ConstraintClass.class).get();
+    assertThat(x1.optInt).isEqualTo(2);
+    assertThat(x1.flag).isFalse();
+    assertThat(x1.optString).isEqualTo("hello world");
+
+    var x2 = config.as("x2", ConstraintClass.class).get();
+    assertThat(x2.optInt).isEqualTo(33);
+
+    assertThatThrownBy(() -> config.as("xf1", ConstraintClass.class).get())
+        .isInstanceOf(CollectedException.class)
+        .hasMessageContaining("is not valid");
+
+    assertThatThrownBy(() -> config.as("xf2", ConstraintClass.class).get())
+        .isInstanceOf(CollectedException.class)
+        .hasMessageContaining("Could not find key");
+
+    var nested = config.as("nested", NestedClass.class).get();
+    assertThat(nested.counter).isEqualTo(5);
+    assertThat(nested.obj.optInt).isEqualTo(33);
+    assertThat(nested.obj.flag).isTrue();
+  }
+
+  private void testSubConf(SqrlConfig config) {
+    assertThat(config.asString("delimited.config.option").get()).isEqualTo("that");
+    assertThat(config.asInt("one").get()).isEqualTo(1);
+    assertThat(config.asString("token").get()).isEqualTo("piff");
+  }
+
+  @SneakyThrows
+  private Path createTempFile() {
+    return Files.createTempFile("configuration", ".json");
+  }
+
   public static class TestObject {
     public String name;
     public int number;
@@ -259,6 +397,43 @@ class SqrlConfigEdgeCasesTest {
       this.name = name;
       this.number = number;
       this.flag = flag;
+    }
+  }
+
+  @AllArgsConstructor
+  @NoArgsConstructor
+  public static class TestClass {
+    int field1;
+
+    @Constraints.MinLength(min = 3)
+    String field2;
+
+    List<String> field3;
+  }
+
+  public static class ConstraintClass {
+    @Constraints.Default int optInt = 33;
+
+    @Constraints.MinLength(min = 5)
+    @Constraints.Default
+    String optString = "x";
+
+    @Constraints.NotNull boolean flag = true;
+  }
+
+  public static class NestedClass {
+    int counter;
+    ConstraintClass obj;
+  }
+
+  public static void testForErrors(Consumer<ErrorCollector> failure) {
+    var errors = ErrorCollector.root();
+    try {
+      failure.accept(errors);
+      fail("");
+    } catch (Exception e) {
+      System.out.println(ErrorPrinter.prettyPrint(errors));
+      assertThat(errors.isFatal()).isTrue();
     }
   }
 }
